apps/python-scraper/requirements.txt
playwright
aiohttp
rapidfuzz
amazon-paapi5
python-dotenv
fastapi
uvicorn[standard]
